requirements.txt
pydantic>=1.10.0
pypdf>=3.0.0
python-pptx>=0.6.21
openpyxl>=3.0.0
